Fix alias origins
Aliases should point at the alias itself. This makes things like

    A.foo

point at the constant literal `A` instead of the location where `class
A` is defined. It also makes things like

    x = A
    x.foo

point to the `x = A` line instead of `class A`

This codepath **only** applies to `isClassOrModule`, so things like
fields, static fields, and type members are unchanged. In particular:

    @x.foo

will point to the place where `@x` is pinned (if it exists)--it is an
error for the field not to exist in `typed: strict`, and even if the
error is silenced it will have been replaced with a harmless local.