Improve error message for "method does not exist" (#3876)
* Thread origin of receiver through DispatchArgs

* guessOverload doesn't actually use fullType arg

* Use origins2Explanations in more places

* Hide origins2Explanations for <self>

Giving the `Cast` instruction a value of `md.loc` gave the weird
impression that the "origin" of `<self>` was either:

- the whole method body
- the whole file contents (for self at the top level)
- the containing ClassDef declLoc

depending on the situation. In almost all of these, the error is going
to be useless:

- showing a whole method body is just going to truncate the method with
  an elipsis for most real-world methods
- same for the whole file contents--makes no sense
- the containing ClassDef declLoc is _maybe_ something the user could
  make sense of, but probably we want to say something more explicit,
  like "we got this type because it was `self`"

In particular, none of those locs would be able to participate in
anything resembling an autocorrect to an assignment in the source code.

Thus, I've removed it.

* Fix alias origins

Aliases should point at the alias itself. This makes things like

    A.foo

point at the constant literal `A` instead of the location where `class
A` is defined. It also makes things like

    x = A
    x.foo

point to the `x = A` line instead of `class A`

This codepath **only** applies to `isClassOrModule`, so things like
fields, static fields, and type members are unchanged. In particular:

    @x.foo

will point to the place where `@x` is pinned (if it exists)--it is an
error for the field not to exist in `typed: strict`, and even if the
error is silenced it will have been replaced with a harmless local.

* Update testdata tests

* Update cli tests

* Update lsp test

* wip fix lsp

* Rebuild

* wip fix lsp