Refactor "Passing a hash ... keyword arguments"
A few lines above this change we see

    auto kwSplatType = Types::approximate(gs, kwSplatArg->type, *constr);

which is to say: `kwSplatType` is the result of taking the existing
`TypeAndOrigins` structure and calling `approximate` on the type part.

For the life of me, I can't construct a test case where this matters.
I tried deleting the call to `approximate` and re-ran the test suite,
and nothing failed.

Nonetheless, I figured "if it ain't broke don't fix it" and so I
re-wrapped up a new `TypeAndOrigins` struct to be able to use `explain`.