Fix other tests
In particular, I think that all error messages for packager names are a
garbage fire right now. We should be suggesting that you might want to
import the other package, or you might have meant to use this similar
thing in another imported package, etc.

So while this error message changes, I don't even think it's worth
considering whether it's "better" or "worse" because they are all bad
errors and they will all have to be fixed before a mass rollout of
packages.