pre-work: Fix bug when combining two basic blocks
We used to say that the `bexit.loc` of the two combined `BasicBlock`'s
was the loc of the `BasicBlock` that unconditionally jumped into the
second block, instead of loc of the second `BasicBlock` jumping out
somewhere else.

We don't actually have anywhere in the codebase that shows this Loc
in an error message until this change.