Only report dead code errors once
Change looks infinitely better ignoring whitespace changes.

The current behavior is that if some piece of code causes one or more
blocks in a chain to become dead, we report a dead code error on the
first instruction inside every block.

With this change, if all the `backEdges` for the current block have
already been marked dead, we don't report an error for this block,
because we assume it means that a dead code error has already been
reported.