Explain where type came from
We have this fancy helper for explaining things, so I figured we may as
well use it.

It's also super handy that we keep all the environments around even
after we're done typechecking a certain block, because technically all
the conditions' variables that I want to be checking are only in scope
in the previous environment.

Also I had to change `getTypeAndOrigin` from private to public because I
didn't want to use `getAndFillTypeAndOrigin` (public), which mutates the
Environment and the VariableUseSite you give it.