Skip dead backedges
This change makes sense intuitively, and I was definitely seeing cases
that this code fixed. I could have sworn that I had a test case for it,
but I can't find one anymore.

Anyways, the intuition is in the comment. I figured it was harmless
enough to include even though the test case it was needed to fix has
been lost to time.