Vastly improve dead code error message (#3929)
* pre-work: Fix bug when combining two basic blocks

We used to say that the `bexit.loc` of the two combined `BasicBlock`'s
was the loc of the `BasicBlock` that unconditionally jumped into the
second block, instead of loc of the second `BasicBlock` jumping out
somewhere else.

We don't actually have anywhere in the codebase that shows this Loc
in an error message until this change.

* Add failing test

* Add ErrorLine on unconditional condition

* Explain where type came from

We have this fancy helper for explaining things, so I figured we may as
well use it.

It's also super handy that we keep all the environments around even
after we're done typechecking a certain block, because technically all
the conditions' variables that I want to be checking are only in scope
in the previous environment.

Also I had to change `getTypeAndOrigin` from private to public because I
didn't want to use `getAndFillTypeAndOrigin` (public), which mutates the
Environment and the VariableUseSite you give it.

* Add autocorrect in --suggest-unsafe mode

This only applies in --suggest-unsafe mode because I figure that's the
common case for dead code errors. It's usually "some codemod made
something untyped into something typed" which causes a bunch of deadcode
errors, because the revealed typing information wasn't actually good
enough.

* Skip dead backedges

This change makes sense intuitively, and I was definitely seeing cases
that this code fixed. I could have sworn that I had a test case for it,
but I can't find one anymore.

Anyways, the intuition is in the comment. I figured it was harmless
enough to include even though the test case it was needed to fix has
been lost to time.

* stog //test/cli:update_suggest-sig