Implement a hash function on every Type
This change will substantially speed up Sorbet start-up in editor on a cold cache (up to ~50%).

When Sorbet starts up on a cold cache, it needs to hash every file in the project. This can take awhile on a large project. Hashing involves running resolver on the file's AST in an empty GlobalState and taking the hash of that GlobalState object.

Metrics show that ~50% of the time spent in `computeFileHash` is in `TypePtr::hash`, which expensively hashes `TypePtr::toString()` (which allocates and formats strings -- with tabs! -- and often calls `Symbol::toString()`).

The new hashing routines use Symbol IDs and should overall be much cheaper to run.