Faster cold cache LSP startup by re-using ASTs for hashing (#3946)
Faster cold cache LSP startup by re-using ASTs for hashing. Hashing is now ~44% faster. That does not mean cold cache startup is 44% faster, as hashing only forms one part of startup. I'm seeing ~15% improvements end-to-end. I have additional plans to make startup time faster.

The speed improvements come from:
* Files are only parsed _once_ at startup. The AST from pipeline::index is rewritten into an AST for an empty GlobalState which can be used for hashing purposes.
* The pass that rewrites the AST does double duty and collects sends and constant references (core::UsageHash), which removes the need for an extra full AST pass.

The main change involved adding a second type of `ast::substitute` pass that supports lazily adding new names to a GlobalState. (The existing pass is designed to re-use a lookup table constructed for all names in a GlobalState for many passes, whereas the lazy pass only cares about the subset of names in one particular AST.)

Measurements show that:
* Hashing the entire project is ~44% faster than before because the cost of the new logic is a small fraction of the cost of reparsing each file. On larger projects, that translates into substantial cold cache startup improvements.
  * With that said, cold cache startup is still appreciably slow; we haven't won yet!
* The new substitute pass is 70% slower than just copying an AST. That is, it's relatively cheap.

The new hashing procedure is not yet used once the LSP server is online. That is, edits by the user still incur two parses. That change is slightly riskier; I dropped a comment explaining why. I plan to tackle that change in the next PR.

In addition to the optimizations, I have added a new CLI parameter called `--force-hashing` that causes sorbet to run hashing over the codebase when run from the CLI. There is no reason why a user would ever need to use this option, but it will make it much easier for me to profile the hashing code.