Type hashing: Avoid NameRef::hash. (#3961)
NameRef::hash uses name IDs in hashing, which are not stable during hashing across file changes; it's impacted by the order in which names are encountered in an AST pass.

NameRef::hash still has use; it is safely used to quickly find a slot in GlobalState's name table. However, we cannot use it in file hashes.

I have replaced the only use of this with a hash of the name's shortname.