Use T::Configuration::AT_LEAST_RUBY_2_7
This handles the remaining call sites.

These call sites are either not as critical or run only at load time, so
it's ok so have these conditional checks where they are.

Specifically, in this change I have added one `if` to the body of
`validate_call`, which is the slow path validator. My benchmarks show
that this is about 3ns of overhead:

    .bind(example).call: 138.425 ns
    .bind_call(example): 96.186 ns
    if AT_LEAST_RUBY_2_7; .bind_call(example); end: 102.712 ns