Re-run rbi/tools/sync-rdoc.rb
I ran this:

    rbi/tools/sync-rdoc.rb rbi/{core,stdlib}/**/*.rbi

but there were a few files that misbehaved, and I ran out of my timebox
trying to figure out why:

- rbi/stdlib/shellwords.rbi
- rbi/stdlib/open3.rbi
- rbi/stdlib/kconv.rbi
- rbi/stdlib/find.rbi
- rbi/stdlib/fileutils.rbi
- rbi/stdlib/etc.rbi
- rbi/stdlib/drb.rbi

In each case, the error was something like this:

    rbi/tools/sync-rdoc.rb:101:in `error!': rbi/stdlib/drb.rbi:316:2: unconsumed sig (RuntimeError)

I'm guessing maybe a few methods were removed? I'm not sure if that's
actually the problem and if so what the best way to work around these
would be.

In any case, I think that this is the first time this script has been
run for a Ruby version different from the version it was built against.
It's probably never been tested for this use case of upgrading a Ruby
version.