Scaffolding to generate call_validation.rb via Bazel (#4019)
* Scaffolding to generate call_validation.rb via Bazel

* format build files

* sto //gems/sorbet-runtime:update_call_validation

* Port style 2.6 to C++

* Change call_validation_2_6.rb by hand

I made this change by manually cutting & pasting the implementation out
of call_validation.rb. You can check this manually by using

    git diff --color-moved

on this commit

* Add conditional require to call_validation.rb

* sto //gems/sorbet-runtime:update_call_validation_2_6_rb

* sto //gems/sorbet-runtime:update_call_validation_2_7_rb

At this point, we're not taking advantage of the --bind-call flag to the
generator script, so this output is identical to the `2_6_rb` command.

* Make use of --bind-call option

* Fix the one test failure!

* rbi: Add UnboundMethod#bind_call

* Use T::Configuration::AT_LEAST_RUBY_2_7

This handles the remaining call sites.

These call sites are either not as critical or run only at load time, so
it's ok so have these conditional checks where they are.

Specifically, in this change I have added one `if` to the body of
`validate_call`, which is the slow path validator. My benchmarks show
that this is about 3ns of overhead:

    .bind(example).call: 138.425 ns
    .bind_call(example): 96.186 ns
    if AT_LEAST_RUBY_2_7; .bind_call(example); end: 102.712 ns

* Add some .bind_call benchmarks to bench:typecheck

* Style

* This isn't actually an UnboundMethod