LSP - Assume large changes cannot take the incremental (fast) path. (#4145)
LSP - Assume large changes cannot take the incremental (fast) path.

To determine if a change (`SorbetWorkspaceEdit`) can be typechecked incrementally, the LSP server does the following:

* Parses the files.
* Calculates each file's file hash.
* Compares each file's hash against the previous version of the file's file hash.

This operation is serial and expensive if many files change. It is also unlikely that a large change can take the fast path. The user impact is that **Sorbet pauses for many seconds and does not respond to user input**.

To fix this problem, this PR does the following: If many files change, it assumes that the edit cannot take the fast path and _defers_ calculating file hashes (which are required for making the fast path determination for subsequent file edits) until just before the change typechecks, when a `WorkerPool` is available to parallelize the work.