Run our lowerings in sequence (#388)
We only have one lowering pass, so it doesn't really matter how it was
scheduled. Put another way: I don't think much thought was every given
to how these passes would compose with each other.

I'm about to add the second lowering pass in #386, and that one *must*
be run entirely after the existing `ClassAndModuleLoading` lowering.

This means that we will traverse the `llvm::Module` once per lowering
pass that we write. This could easily become slow.

That being said, there are only two passes right now, and so far all of
our work has been in IREmitter itself, with not much indication that
we'll be adding a lot more lowerings.

When we decide that it's time to start optimizing compile-time
performance, we can revisit this decision. At the very least, it should
be easy enough to schedule `ClassAndModuleLoading` and the future
lowering in #386 (`ObjIsKindOf`) to run in sequence **at each call
site**, but still in one `llvm::ModulePass`.