`FixedHash` serialization should not assume symbol keys
`FixedHash` current serializes itself assuming that the keys are symbols
but it can also be the case that keys are strings. Thus, it is more
correct to use the `{ a => b }` syntax when serializing fixed hashes
entries with non-symbol keys, along with calling `inspect` on those keys
instead of an implicit `to_s`.