Implement name-based intrinsic for `nil?` (#440)
* Implement name-based intrinsic for `nil?`

We have taken the approach that we should make `.nil?` work everywhere
and always mean the expected thing, rather than allowing flexibility to
redefine it, or to raise a `NoMethodError` when called on `BasicObject`
instances.

We have changed Stripe's codebase to introduce this monkey patch:

<http://go/ps/pull/332445>

and we have also verified that adding such a monkey patch does not
adversly affect the interpreted code (the VM's fast path checks still
trigger).

Timings:

| source                     | interpreted | compiled |
| ---                        | ---         | ---      |
| stripe/while_10_000_000.rb | 0.187s      | 0.067s   |
| stripe/nil_p.rb            | 0.424s      | 0.067s   |
| stripe/nil_p.rb - baseline | 0.237s      | 0.000s   |

LLVM is able to completely eliminate the body of the while loop.
I considered whether to change the benchmark, but I considered this a
feature not a bug, so I figured it was worth keeping the benchmark as
written.

We also tried writing this intrinsic a few other ways. Timing
information for those approaches:

| source                                              | interpreted | compiled |
| ------                                              | ----------- | -------- |
| stripe/nil_p.rb - baseline @ master                 | 0.230s      | 0.920s   |
| stripe/nil_p.rb - baseline @ symbol based intrinsic | 0.237s      | 0.342s   |
| stripe/nil_p.rb - baseline @ inline cache intrinsic | 0.233s      | 0.259s   |

(In the above table, `interpreted` are all the same. Timing differences
for interpreted indicate fluctuations in measurement.)

Specifically:

- The symbol-based intrinsic spent nearly all of it's time climbing back
  up the inheritance hierarchy to find `Kernel` close to the very top.
- The inline inline cache based intrinsic (WIP in #421) was comparable
  in performance (much better than the compiler), but still required
  doing a VM method search, which we'd like to not have to do.

* Delete word