Move some logic from ErrorFlusherStdout to AutocorrectSuggestion
This makes it easier to use just the parts I want in the test.

Conceptually, it splits the steps into "read files and apply
autocorrects" and "write autocorrected files out". In the tests, I just
want the reading--the writing will happen in a special way so as to not
clobber the input files.

A note on the deleted test:

I don't like Sorbet's unit tests.

This one in particular is completely redundant with Sorbet's existing
autocorrect tests.

To keep this test, I would have had to rewrite the test to operate with
`FileSystem &fs`. But that's exactly the code that the new test suite
I'm writing does, so why bother?