Add `*.autocorrects.exp` expectation tests (#4231)
* Refactor a private Loc helper into a public File helper

Same reasoning as previous commit: I want to be able to use this in the
tests.

* Move some logic from ErrorFlusherStdout to AutocorrectSuggestion

This makes it easier to use just the parts I want in the test.

Conceptually, it splits the steps into "read files and apply
autocorrects" and "write autocorrected files out". In the tests, I just
want the reading--the writing will happen in a special way so as to not
clobber the input files.

A note on the deleted test:

I don't like Sorbet's unit tests.

This one in particular is completely redundant with Sorbet's existing
autocorrect tests.

To keep this test, I would have had to rewrite the test to operate with
`FileSystem &fs`. But that's exactly the code that the new test suite
I'm writing does, so why bother?

* Check for *.autocorrects.exp expectation files

* Incorporate autocorrects.exp into update_testdata_exp.sh

* Add example for autocorrect_method_return.rb

* Add an example for multi-file support too

* Update README.md to talk about multi file exp tests

* Allow enabling --suggest-unsafe in testdata tests

* Add test showcasing --suggest-unsafe