LSP: Limit maximum number of diagnostics to report to the editor. (#4239)
Limit the maximum number of diagnostics to report to the editor when in LSP mode. I have set the value to `1000`, but I may ratchet it down with testing on Stripe's codebase. This limit does not apply in CLI mode.

Before explaining how it works, recall that Sorbet's LSP server uses the concept of an epoch. An epoch is simply a typecheck run, which can be full (i.e., retypecheck _every_ file) or incremental (i.e., only retypecheck a subset of the files in the project). Incremental typecheck runs can _preempt_ full typecheck runs to respond to code edits during a full typecheck, so multiple epochs can be in play at once.

To manage the complexity of error reporting in the face of preemption, Sorbet's LSP server uses an `ErrorReporter` class which is responsible for sending diagnostics (i.e., typechecking errors) to the client. It does some sanity checking to avoid sending stale typechecking errors, since it is possible that a full typecheck may report errors for an older file revision _after_ it is preempted by a incremental typecheck. 

One other important detail of the LSP protocol is that servers update the set of errors in a file by sending the full list of errors in the file. That is, if a file has 5 errors, it is not possible to send 2 and then send 3 -- the server must send a complete listing. A corollary of this design decision is if a file has 0 errors but previously had >0, an empty list of diagnostics must be sent to the editor to invalidate the old errors.

With this PR, `ErrorReporter` is also responsible for limiting the number of diagnostics sent to the editor in the following manner:
* `ErrorReporter` contains a counter of the number of errors reported to the client (`clientErrorCount`).
* `clientErrorCount` is kept up-to-date as new errors are reported for files -- including when an incremental typechecking run shrinks the error count.
* `clientErrorCount` is reset to `0` whenever a full typecheck runs since that typecheck will re-report all errors in the project.

### Limitations

#### Determinism

When the limit is enforced, the set of errors reported is not guaranteed to be consistent between runs, since errors are reported in the order received from worker threads. Thus, the user may see a slightly different set of errors with each full retypecheck. Fixing this problem is somewhat at odds with error streaming.

One possible solution that could be bolted on to this effort is to stream out errors in a specific file order rather than in the order typechecked. Files enter the worker queue in a deterministic order, and we could re-sort the files on the way out into the same order. This fix would slightly reduce the responsiveness of error streaming and may not be worth the complication.


#### UX

The user does not receive an indication that Sorbet is limiting errors. `clangd` limits errors and also does not provide any user indication that it is doing so. I contemplated adding a faux error to a file that stated that errors are limited, but that is easily missed. Suggestions appreciated.

### Motivation
<!-- Why make this change? Describe the problem, not the solution. This can also be a link to an issue. -->

Sorbet effectively crashes VS Code when it reports thousands of errors.

Fixes https://github.com/sorbet/sorbet/issues/3604.