Prohibit calls on T.class_of, fix T.class_of in generics (#4340)
* Add failing test

* Prohibit calls on T.class_of

* Fix failing test

* Fixup existing tests

The new error is basically an artifact of the existing error, but one is
raised by type_syntax.cc and the other is raised here in calls.cc.

I like the new error, I think it makes sense, and it's basically the
same error you'd get if you'd written `T.class_of(1)`. That consistency
leads me to believe that the implementation based on `unwrapType` and
then `unwrapSymbol` is robust and general.

* CFG exp file changes

* symbol-table exp changes

Another error snuck into this file, so the line numbers changed.