Add some failing tests
The `is_a?` test fails in the same way on `master` already, which makes
me ok with the fact that the same bug already exists for `Module#<` and
will exist for `Module#<=`.

There is a ticket open to potentially follow up with a fix for that:

<https://github.com/sorbet/sorbet/issues/4358>

In the mean time, there is only one problem due to this bug on Stripe's
codebase after this change, which seems fine to go ahead with.