Change type caching to stop polluting modules
Fixes #4391

Since #3309 was merged, whenever a module is coerced to a Sorbet type
an instance variable is created on the original module to cache the
created `T::Types::Simple` instance.

However, that action ends up adding instance variables to types not
under Sorbet runtime's control. Even worse, the value that is added has
a reference to the module, so any operation applied to the parent module
that walks over instance variables recursively would end up in an
infinite recursion. This is already happening in #4391 for JSON
serialization via ActiveSupport.

The fix is to store the cached values on a `WeakMap` on the `Pool`
class. This ensures that the cached values can be garbage collected, stops
polluting modules and does not seem to be negatively impacting the
performance either.