Don't do the clever thing to fix ENFORCE
I thought I could get away with typing less code, but it tripped an
ENFORCE in `dupType`. I think we could probably just delete the ENFORCE,
because I don't think the property it's ENFORCE'ing is actually relied
on or a particularly good invariant to maintain, but I figure it's easy
enough to leave it as is and change my code.