Make DispatchArgs::suppressErrors required
Errors were not always being suppressed properly, because certain
locations forgot to propagate the old value when constructing a new
`DispatchArgs` struct.

This meant that we were actually doing substantially more work than we
meant to.

This might also fix a source of non-deterministic behavior in Sorbet on
Stripe's codebase (not yet tested) but the performance benefits (not yet
measured) should be justification enough to land this.