Move getSubclasses to LSP-specific helper (#4703)
* Move getSubclasses to LSP-specific helper

`getSubclasses` is a very slow operation, and does not belong outside of
LSP, because it is not a primitive that should ever be used in Sorbet's
core type checking pipeline.

History: this started out as a helper function in an anonymous namespace
in rename.cc. `workspace/rename` is fine with having the rename request
take a little bit, because it more than makes up for how much time the
user would have had to spend renaming things manually.

The helper was moved under Symbol to be shared by
`textDocumement/implementation`. That was overly eager--it should have
been factored to an LSP-specific location instead.

I've also added a comment to capture these considerations for the
future.

* forgot one