Decrement depth in `flatten` only if flattenning arrays
The `recursivelyFlattenArrays` method was unconditionally decrementing
the `depth` counter before performing any operations, but that leads to
wrong results.

We should only be decrementing the `depth` counter if we are descending
into the member type of an array or tuple type.

For example, the fact that we were acting like we were flattening one
level when trying to resolve the flatten result for left and right of an
`Or` type was producing the wrong result for this:

```ruby
[
  case rand
  when (0..0.5)
    1
  else
    ["2"]
  end
].flatten(1)
```

The code is clearly equivalent to `[1].flatten(1) #=> [1]` or
`[["2"]].flatten(1) #=> ["2"]`, and, thus should have the type:
`T::Array[T.any(Integer, String)]`.

However, since we were doing the decrement eagerly, we were getting
`T::Array[T.any(T::Array[String], Integer)]`.