Reimplement `flatMapHack` as `map.flatten(1)` (#4722)
The current implementation of `flat_map` was using code which was
similar to the `flatten` code but it was incomplete since it was not
respecting the return type of `to_ary` methods implemented by the types
inside arrays.

A better implementation is to implement it as `map.flatten(1)`, that is
dispatch `flatten(1)` on the return type of the `map`ping operation,
which allows us to reuse the full `Array_flatten` implementation that
uses `to_ary` return types for better typing. Moreover, any improvements
to `Array_flatten` will be reflected automatically in `flat_map`
implementation as well.