Allow passing arbitrary command line args in sorbet.run
This does not attempt to pass any of those options to LSP at the moment.
That can cause surprises, for example, if you pass a flag that affects
type checking (like to enable requires ancestor) or that affects LSP,
those flags won't have an effect.

(This is because of an annoying implementation detail, which is that we
do not actually pretend to launch Sorbet with command line arguments in
service of LSP in the browser, preferring instead to use the LSP wrapper
that we've built for running our assorted test harnesses, all of which
assume that you're passing in a global state and parsed set of args.
That's not to say it would be impossible to invoke the arg parsing logic
and pass them in, it's just work I don't want to do right now.)

This change must be merged and deployed to sorbet.run before the
corresponding change upstream. This PR implements a pseudo-backwards
incompatible change: it changes the meaning of the `typecheck` function
from accepting a Ruby program as a string to accepting a stringified
JSON array of command line args.

In the interim period where the WASM payload has been updated on
sorbet.run but sorbet.run itself hasn't been updated, a warning will be
printed in the `#output` div on the page.

We will follow up quickly with a change to update the client to know
to send JSON instead of a Ruby program, so that warning should only show
for a few moments.