Better error message for sealed with grandchild (#4795)
It's not clear to me whether this should be allowed or not. This is
currently allowed statically. But at least for the moment, I figure I'd
add a better error message so people can track down the cause. We might
end up relaxing this but I haven't thought about it.