Refine error message when constant names are the same (#4757)
If we have the name of the expected type and the name of the object
class the same but referring to different constants, the default error
message ends up being confusing. In those circumstances, it would be
helpful to show that the constants refer to different objects even
though they have the same name and that this is probably a code
reloading issue.