Add `Pathname#glob` to stdlib. Fix second argument to both glob variants - not a String but Integer.
Previously, Sorbet knew only about `Pathname.glob` the class-method. It is nearly identical,
but globs from PWD instead.

The second argument to both (and to Dir.glob which is used underneath) is a logical-OR of flags
defined in File, e.g. File::FNM_DOTMATCH which makes `*` match filenames beginning with a dot.