.ruby-version: 2.6.3 -> 2.7.2
We already run the sorbet-runtime tests in 2.6.3 and 2.7.2. This commit
doesn't change that. This is more of a convenience, for e.g., when I run
`irb` from the `~/sorbet` folder, or I have ALE use `ruby` to check a
program for syntax errors, and the file is using 2.7-only syntax.