use a faster output format for ragel state machines
Ragel has several different output formats for the state machines it
produces.  The default corresponds to the `-T0` option, and is described as
"binary search table-driven" in the Ragel user manual.

Profiling with `valgrind --tool=callgrind` indicates that a significant
amount of time is spent in the binary search in the generated code: 10% or
more of the total parse time for a large Ruby
file (`ruby/lib/rdoc/markdown.rb`).

Experimenting with other output options revealed `-G1`, which uses `goto` to
manage the state machine and features less code reuse between different
actions, produced much faster code: 30% fewer instructions executed during
parsing according to Callgrind.

Experimenting with this new option on Stripe's codebase (using
`--stop-after=parser` and an uncached build to force the parser to execute)
reduced the max parse time by ~20% and the wall-clock time for the
`--stop-after=parser` run by approximately the same percentage.

The binary size for a `--config=release-linux` build increases by about
300kb, which seems like a reasonable tradeoff.