Change line splitting test
It looks like this test was trying to show that multi-line locs in error
messages get the little `...`. It so happened that the error it was
using to demonstrate that behavior no longer will do that (which is
good!). I left the existing test to capture the improvement, and also
added another example of a test where this will still happen (so that we
can catch any future changes that somehow regress multi-line loc error
printing).