Recover better from "Dynamic constant ref" errors
Our current pipeline would basically entirely ignore the dynamic scopes
for code like

    x.foo::BAR

That's usually fine for Sorbet (Sorbet codebases have to accept this
limitation of Sorbet), except that it can happen as a result of
a partially written Ruby file, e.g.

    x.
    Foo::Bar

Our parser sees this as `x.Foo()::Bar` as is (without extra error
recovery), and in particular we would like to be able to see the `x.` as
the start of a method call whose method name hasn't been provided yet.

The solution is to make sure that even if we aren't going to attempt to
ingest the `x.Foo()` for the purpose of constant resolution, we should
at least let is fall through infer so that `x` gets assigned a type.

A note about the implementation below: it appears that our TreeMap
abstraction only allows post-transforms on `UnresolvedConstantLit`, and
doesn't recursively map over the `scope` field. I noticed that
`Substitute.cc` gets around this by just doing the `TreeMap::apply`
directly, which works for us too. My first attempt was to build out
`TreeMap` to recurse over the scopes, but this seemed less invasive.