Delete most TODOs
applySig: we forward the `funLoc` when rewriting `sig` to
`Sorbet::Private::Static.sig` so that's fine.

The `id->name` case is turning a local into a send, which can use the
identifier's loc as the method loc.

Not storing `methodLoc` for `!x`, because it won't work well in any of
the other places where we'd want to rely on this (method rename,
completion, etc.)--in all those cases you'd have to special case `!x`
handling, even if you had the method loc.