Track location for method name specifically (#5104)
* wip: Add methodLoc to parser::{Send,CSend}

* wip: Add funLoc to ast::Send

* Reorder cfg::Send fields

* wip: Add funLoc to cfg::Send

* wip: Add fun to core::CallLocs

* Use funLoc in calls.cc

* Factor out Loc::empty()

* Don't set funLoc for [] and []=

* Some error locations change

* Change line splitting test

It looks like this test was trying to show that multi-line locs in error
messages get the little `...`. It so happened that the error it was
using to demonstrate that behavior no longer will do that (which is
good!). I left the existing test to capture the improvement, and also
added another example of a test where this will still happen (so that we
can catch any future changes that somehow regress multi-line loc error
printing).

* Update cli tests

* Fix bug in parse-tree-json-with-locs

This format looks entirely broken: the locs aren't actually there--only
the `declLoc` (and now `methodLoc` fields are).

* Leave a TODO for thought

* update lsp test

* Delete most TODOs

applySig: we forward the `funLoc` when rewriting `sig` to
`Sorbet::Private::Static.sig` so that's fine.

The `id->name` case is turning a local into a send, which can use the
identifier's loc as the method loc.

Not storing `methodLoc` for `!x`, because it won't work well in any of
the other places where we'd want to rely on this (method rename,
completion, etc.)--in all those cases you'd have to special case `!x`
handling, even if you had the method loc.