Better method completion given syntax errors (#5081)
* lexer changes

* no-op: Debug printers for some tokens

* Recover from some bad method calls

* Record improvements in tests

* Add inference test (before)

* Don't leak parser implementation in type errors

* Implement support for `x.` completion

* Fix AddressSanitizer failures

Whoops, `ts` is "token start" which is basically the start of the last
token it read when it was executing the scanner actions and/or doing
backtracking, not the start of the whole buffer, which is what the
`start` and `end` tokens are relative to.

* Change keyword_method_names test

This test is just meant to capture the fact that we're already deviating
with what the Ruby parser would do. We're fine with these changes

The only potentially concerning thing is that to get the test to
somewhat still match up withh the old behavior, we manually added an
extra `end` keyword. The justification for why that's like that is that
it's a separate ticket to track mismatched `end` keywords (for def,
begin, etc.) that we'll get to eventually (maybe even shortly).

* fixup redo test assertions

* Parser traces now show location information

* Add a lot of tests

* Add test for conditional send

* code review fixups

* Fix semantic conflicts after merging unrelated branches

* Empty completion prefix when outside response's funLoc

* Recover better from "Dynamic constant ref" errors

Our current pipeline would basically entirely ignore the dynamic scopes
for code like

    x.foo::BAR

That's usually fine for Sorbet (Sorbet codebases have to accept this
limitation of Sorbet), except that it can happen as a result of
a partially written Ruby file, e.g.

    x.
    Foo::Bar

Our parser sees this as `x.Foo()::Bar` as is (without extra error
recovery), and in particular we would like to be able to see the `x.` as
the start of a method call whose method name hasn't been provided yet.

The solution is to make sure that even if we aren't going to attempt to
ingest the `x.Foo()` for the purpose of constant resolution, we should
at least let is fall through infer so that `x` gets assigned a type.

A note about the implementation below: it appears that our TreeMap
abstraction only allows post-transforms on `UnresolvedConstantLit`, and
doesn't recursively map over the `scope` field. I noticed that
`Substitute.cc` gets around this by just doing the `TreeMap::apply`
directly, which works for us too. My first attempt was to build out
`TreeMap` to recurse over the scopes, but this seemed less invasive.