Use overloading for `show()` (#5153)
This makes it easier to use `.show` in the debugger. LLDB doesn't know
about default arguments in header files, so it thinks that two args are
required. Since `show` is so commonly used for debugging things, it's
nice to keep it short and not introduce extra "required" args.

Overloading makes it work, because it actually shows up as two symbols
in the object file.