Showcase multiple "end of file" errors
This is an artifact of how it works right now: every time the
kend_or_eof rule reduces, it'll add another error. So if it reduces
multiple times because there are multiple `kEND` tokens missing, each
one will get a new error. It's kind of weird, and we could attempt to
dedup, but honestly it seems kind of not terrible.