Thread line start offset through parser
This will let us compute the indentation of the line a token is on.

We might want to populate the indentation of the token on construction,
but for simplicitly right now I'm just going to ask for the indentation
at the points where I ask for it.

This might not be the most efficent path forwards, but it might be the
sort of thing that we can just elide/not populate except once we know
there's a syntax error, and we can afford to be a little slower than
usual.

Also, this bloats the size of a `token` by one `size_t`.