Indentation-aware parser error recovery (#5000)
* no-op: Factor out makeDriver helper

We're going to be making multiple drivers. It'll help to have a helper
function for that.

* Thread line start offset through parser

This will let us compute the indentation of the line a token is on.

We might want to populate the indentation of the token on construction,
but for simplicitly right now I'm just going to ask for the indentation
at the points where I ask for it.

This might not be the most efficent path forwards, but it might be the
sort of thing that we can just elide/not populate except once we know
there's a syntax error, and we can afford to be a little slower than
usual.

Also, this bloats the size of a `token` by one `size_t`.

* Add compare_indent_level helper

* Thread `indendationAware` boolean through driver

* Add rewind_if_dedented helper

This uses the new indentationAware boolean and the compare_indent_level
helper to make a decision about whether to reduce normally or whether to
emit an error that the `end` keyword indentation was mismatched, and
rewind the lexer to re-lex the `end` keyword if so.

* Existing test improves

* Rename a test

* Show behavior before this PR

* Show behavior after this PR

* Also update newline_s after line comment

* Add new completion tests

* Remove TODO

* Add another test

* whoops, recorded this test wrong

* Turns out there's an overload

* Use addErrorNote to explain the message a bit

* Report error recovery errors as separate code

... so that we can write a section for them in the error docs.

* Prefix with `Hint:`

* Delete ErrorToError, just use functions

This change looks better ignoring whitespace.

* Always report original errors

This way we can make the messaging around the hint errors more clear
("they're just hints, might be wrong")

* Add one more line of context to the error

* Fix wording in error-reference

* whoops double loc

* Suggested refactoring

* reorder if/else

* Remove errorToError

* Fix comment

* Add cli test showing error lines

* Change error message and location

* Add a Timer for the withIndentationAware phase

Here's hoping that this always takes so little time that it never shows
up in the log.